UPSTREAM: amd/stoneyridge: Add USB ASL for D0/D3cold
Add methods, and call them, for transitioning EHCI and xHCI to D0 or
D3cold.  Add device objects necessary for waking the system via USB.

In order for USB to wake the system, it must be in the D3cold state.
Then on resume, its firmware must be reloaded.

This code relies heavily on AMD's FchCarrizo.asl (delivered in NDA PI
package), and has been modified to fit the coreboot ASL names.  In
addition, AMD's methodology is to generate a SW SMI for saving/restoring
certain settings.  This has been ported into U3D0 and U3D3, as the
necessary registers are now publicly documented.

BUG=b:77602074

Change-Id: I081c9e4ac5fab1f448881ee65e1171fc24884141
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: fdb846ddf270e5a3406a835fb67a5c1b049f9bf3
Original-Change-Id: I83d0dce13411601691318cc67c99adf291ccf3bb
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/28772
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Reviewed-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1254798