UPSTREAM: amd/stoneyridge: Add USB settings to gnvs
A later patch will rely on two USB settings from the BIOS.  Add these
to the global_gnvs_t structure.

The first is a data that will be used to locate the xHCI firmware for
reloading after a resume.  Although the existing calculations will be
somewhat simple, keeping this on the coreboot side will help in the
event multiple FWs are eventually in the build.

The second item is a usable EHCI base address that may be programmed
during S3 suspend and resume.  At the time the PTS and WAK code runs,
the BAR will be clear.

BUG=b:77602074

Change-Id: I75d0b44d7833750002ac1d47503fa8d4935cde7b
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: d61e347bffae29e6db179641652602226711e401
Original-Change-Id: I32205ac8a6908cca4a38dd68a7c7b591e76c06bb
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/28766
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1254792