UPSTREAM: amd/stoneyridge: Remove SATA D0 on suspend
Remove the step of setting the SATA controller to S0 as the system is
entering S3.  This had been duplicated from AMD's FchCarrizo.asl file,
but upon closer inspection, the conditions for this step to run cannot
be met.  This does not affect Grunt's behavior, as the SATA controller
is disabled.

TEST=Suspend and resume Grunt
BUG=b:77602074

Change-Id: I878a4e010b9f37f8ee2f49655cbcd180d9403345
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: 0425b0a4bbd964bda9368aa9adc40a71d0b83362
Original-Change-Id: Ib269a5363d03c7048abd0c8a9a28df92a773790c
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/28863
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1263088
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>