UPSTREAM: libpayload: gdb: Factor out gdb_handle_reentrant_exception() from arm32
The arm32 GDB architecture code contains a little hack that allows it to
(sort of) correctly deal with a reentrant exception triggered from
within the GDB stub. The main logic for this isn't really arm32 specific
and could be useful for other architectures as well, so factor it out
into a separate function.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ie4ac9d0c9a3c419a7796bac315e6a30216fce228
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: 1bfda7293ac0f06681097a8cd6964d86af0b24ba
Original-Change-Id: I3c6db8cecf1e86bba23de6fd2ac9fdf0cf69d3c6
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/29019
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1280004
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>