UPSTREAM: soc/amd/stoneyridge: Define PM USB Enable register
Make #define definitions for PMxEF and replace the hardcoded values.

Note that this doesn't change the current functionality of the source.
The existing code has been propogated from the sb//hudson port, which
seems to attempt to enable 100% of all OHCI and EHCI controllers that
may be present in the system.

BUG=none
BRANCH=none
TEST=none

Change-Id: I39b8c2a3dd4a847c7761bc149b3052fd2092fa2c
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: d453081a57508e25f6d46223e1e8a0f6f93c85ce
Original-Change-Id: I6018b0062730de19e3283a010144dfedc2b11423
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/29075
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1280009
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>