UPSTREAM: ec/google/common: Add a common MEC interface
In order to re-use the MEC interface code in the Chrome EC driver
move it to a common directory within the ec/google directory.

The Chrome EC driver itself is changed to use this interface in the
next commit, and future commits will introduce a new EC that also
uses this interface.

BUG=none
BRANCH=none
TEST=none

Change-Id: I7a888adc5a7dcdc691143bb6ff7573b6bc41aed1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 26cf00ab498a78b59c9170d38d6ee7ced930259c
Original-Change-Id: I13516b5e4c4c49f53bb998366284a26703142e2a
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/29111
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1288755
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>