UPSTREAM: ec/google/wilco: Add a bootblock function for early init
Add a function for use in bootblock stage that performs early init
of the EC, in particular setting it up for UART passthrough so a
legacy serial port can be used by the host.

This needs to be called by the mainboard that intends to use it
in bootblock in order for the UART to be available in later stages.

Some of the PNP style programming may look odd, but it is following
the EC specification which is not entirely standard.  This code has
been tested on a board with this EC and it is functional.

BUG=none
BRANCH=none
TEST=none

Change-Id: I04d0c766c3aac1f5eec6b4c4a81f97de05d8a2f3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ea98b40efe0ae6d514ebc1d797e564abe79ea34d
Original-Change-Id: I9d6935a9fdf0d7290a94bf2ee565ef2a7c00ecc7
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/29121
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1317468
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>