UPSTREAM: mb/google/poppy: add the smi_events back
Before entering the OS, the AP relies on the smi handler to shutdown the
system when the lid closes. Without the smi_events setting, the AP will
not receive the smi handler. As a result, the AP won't shutdown and will
always keep in S0.
This problem is caused by the https://review.coreboot.org/c/coreboot/+/28983
and this patch adds the smi_events back to support the smi handler for
the lid close.

BRANCH=master
BUG=b:115572596
TEST=test_that -b ${BOARD} ${IP} firmware_ECLidShutdown

Change-Id: I27bea31b4578a4fe4943e02a9d5a3bce2da11bf7
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 2ce56f152367de3b717658a1e848c4c1da2e23f0
Original-Change-Id: Id82311a8ccd109f9c26516f59a45bdf34da98529
Original-Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/29191
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1297498
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>