UPSTREAM: soc/amd/stoneyridge: Get rid of void pointer math
Pointer math with void pointers is illegal in many compilers, though it
works with GCC because it assumes size of void to be 1. Change the pointers
or add parenthesis to force a proper order that will not cause compile
errors if compiled with a different compiler, and more importantly, don't
have unsuspected side effects.

BUG=b:118484178
TEST=Build and boot grunt.

Change-Id: I5a779fd149d38dc901fe085540e3d8e6acf7dbc1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9f25e9d4809fe41551243b679ac1ff8b36067b5f
Original-Change-Id: Ibfeb83893f09cb897d459856aff2a4ab2a74e6e5
Original-Signed-off-by: Richard Spiegel <richard.spiegel@silverbackltd.com>
Original-Reviewed-on: https://review.coreboot.org/29337
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/1317805
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>