UPSTREAM: Add selfboot to the romstage
Now that bounce buffers are gone, and we can elide the selfload checking
code at build time, it is safe to add selfboot to the romstage.

Make it so.

This required a few other tweaks to rules.h and selfboot.c
to make it buildon ARM.

BUG=none
BRANCH=none
TEST=none

Change-Id: Icf8aeb533d2946c6abca6ebd21b2c370bf4d532b
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: 3d8e53f85ad4cf6547f3e4c408aa49460fd76eec
Original-Change-Id: Ib6540921ad7cd7d58bfeab881d3978325b303cc2
Original-Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/29338
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Philipp Hug <philipp@hug.cx>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1308589
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Reviewed-by: ron minnich <rminnich@google.com>