UPSTREAM: nb/intel/haswell/gma: Support boards that have DDI E connected
On an ASRock H81M-HDS neither libgfxinit, nor Linux, is able to
initialise the display when lanes are not configured to be shared
between DDI A and DDI E.

Intel's reference manual [1] states that the decision to share lanes
between DDI A and DDI E is "based on board configuration". Hence, add a
new field to the devicetree that boards can set. All existing Haswell
boards have this unset, thus taking a value of 0, so there is no change
to existing behaviour.

[1]: Intel Open Source Graphics Programmer's Reference Manual (PRM)
     Volume 2c: Command Reference: Registers (Haswell)
     https://01.org/linuxgraphics/documentation/hardware-specification-prms/2013-intel-core-processor-family

BUG=none
BRANCH=none
TEST=none

Change-Id: I7fbee9fbf27665efbdf6834c6b8b10edd72196a5
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1a73eb08e79b4a53702ff9e9103bb3352391892b
Original-Change-Id: I6f7832293215d2b53e31b0a5c985e6098eb72f1b
Original-Signed-off-by: Tristan Corrick <tristan@corrick.kiwi>
Original-Reviewed-on: https://review.coreboot.org/29385
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1317489
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>