UPSTREAM: mainboard: Add ASRock H81M-HDS
Tested with GRUB 2.02 as a payload, booting Debian GNU/Linux 9.5 with
kernel 4.9.

This board works quite well under coreboot. A list of what works and
what doesn't can be found in the documentation part of this commit.

The file `data.vbt` matches the VBT in the latest stable version of the
vendor firmware (version 2.20).

BUG=none
BRANCH=none
TEST=none

Change-Id: I8f789848919bd7b413c04aafa8b4f5006e83a65d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 3693294112a717aeb3f14db80f9e753dd376ca58
Original-Change-Id: I53483bb9fa335e86e85dfc487fef03fce4b85e2a
Original-Signed-off-by: Tristan Corrick <tristan@corrick.kiwi>
Original-Reviewed-on: https://review.coreboot.org/29390
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/1341389
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>