UPSTREAM: mb/google/sarien: Add new mainboard
Sarien is a new board using Intel Whiskey Lake SOC.  It also uses
the newly added Wilco EC, enabled in a separate commit.

Sarien is not a true reference board, it is just one variant of
a very similar design.  For that reason it is not considered the
baseboard but rather a standalone variant.

BUG=none
BRANCH=none
TEST=none

Change-Id: I78e65e1adc50bc03cb12a141f7de9a921f74df2a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 558602ff4043082f3e0fc91cbf4905302853a94c
Original-Change-Id: I2e38f617694ed2c2ef746ff8083f2bfd58cbc775
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/29409
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1317500
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>