UPSTREAM: coreboot-sdk: Prefer gnat package over gnat-6
The choice of `gnat-6` was originally an optimization because the meta-
package `gnat` installs not only the current GNAT version but also other
unwanted (and hard to explain) dependencies. Later it was necessary
because GCC 8 couldn't compile our older crossgcc.

Now that we switched crossgcc to GCC 8.1, `gnat` should be fine.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib6ec64eef1b6f851c83af0bd6533e9408ed403df
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b133c61a9151b3b7b9c6b62091a606498c5f7568
Original-Change-Id: Ica8a1f9d6d71a74ffc4ec76aa0cfbe4b604cde1b
Original-Signed-off-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-on: https://review.coreboot.org/29454
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1318934
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>