UPSTREAM: soc/intel/common/block/acpi: fix P-States extra entry
The ratio_max step is appearing twice when (ratio_max - ratio_min)
is evenly divisible by the ratio step.

This is because in this case there are no rounding down of ratio_max in
the for loop.

Thanks Jay Talbott for the step calculation algorithm.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id160a3a388e01accb2a2991e97b4b5175cd5ddcc
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: c2540a958b5e8f4ad4ea0db24c408387ae1cb5f8
Original-Change-Id: I91090b4d87eb82b57055c24271d679d1cbb3b7a7
Original-Signed-off-by: Julien Viard de Galbert <jviarddegalbert@online.net>
Original-Reviewed-on: https://review.coreboot.org/c/25429
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1434506
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>