UPSTREAM: drivers/uart/Kconfig: Be smarter about DRIVERS_UART_8250IO
It defaults to y to avoid having to select it per mainboard. But that
makes a mess because it results in linker conflicts unless other UART
drivers disable it explicitly.

We try to be smarter about the default value for now. The real solu-
tion would be to hardcode it per mainboard. But who knows which boards
actually have it?

BUG=none
BRANCH=none
TEST=none

Change-Id: I812bcb582211ff0f3bdb3f11193d31ac87e35ecf
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 3eb720c36e55948c60e158129721705628a9c249
Original-Change-Id: I7e755fe0e4f6d1c31ef2854603a5510c3cdc4967
Original-Signed-off-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-on: https://review.coreboot.org/c/29571
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1346309
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>